Asserções (Assert)

+++ TestesDeSoftware/01 - Testes de Unidade/Demo/Calculadora.cs
@@ -11,7 +11,7 @@ namespace Demo
             return v1 + v2;
         }
 
-        public double Dividir(double v1, double v2)
+        public double Dividir(int v1, int v2)
         {
             return v1 / v2;
         }
